bam/bai/index/reference_sequence: Fix clippy warnings
This fixes `clippy::iter_kv_map` and `clippy::unnecessary_cast`.

These were added/changed in Rust 1.66.0 / clippy 0.1.66 (69f9c33 2022-12-12).

+++ noodles-bam/src/bai/index/reference_sequence.rs
@@ -101,11 +101,7 @@ impl ReferenceSequence {
         let (start, end) = resolve_interval(interval)?;
         let region_bins = region_to_bins(start, end);
 
-        let query_bins = self
-            .bins()
-            .iter()
-            .filter(|b| region_bins[b.id() as usize])
-            .collect();
+        let query_bins = self.bins().iter().filter(|b| region_bins[b.id()]).collect();
 
         Ok(query_bins)
     }
@@ -191,7 +187,7 @@ fn region_to_bins(start: Position, end: Position) -> BitVec {
     let start = usize::from(start) - 1;
     let end = usize::from(end) - 1;
 
-    let mut bins = BitVec::from_elem(bin::MAX_ID as usize, false);
+    let mut bins = BitVec::from_elem(bin::MAX_ID, false);
     bins.set(0, true);
 
     for k in (1 + (start >> 26))..=(1 + (end >> 26)) {
@@ -240,7 +236,7 @@ mod tests {
         let start = Position::try_from(8)?;
         let end = Position::try_from(13)?;
         let actual = region_to_bins(start, end);
-        let mut expected = BitVec::from_elem(bin::MAX_ID as usize, false);
+        let mut expected = BitVec::from_elem(bin::MAX_ID, false);
         for &k in &[0, 1, 9, 73, 585, 4681] {
             expected.set(k, true);
         }
@@ -249,7 +245,7 @@ mod tests {
         let start = Position::try_from(63245985)?;
         let end = Position::try_from(63255986)?;
         let actual = region_to_bins(start, end);
-        let mut expected = BitVec::from_elem(bin::MAX_ID as usize, false);
+        let mut expected = BitVec::from_elem(bin::MAX_ID, false);
         for &k in &[0, 1, 16, 133, 1067, 8541] {
             expected.set(k, true);
         }
